update status bar
