+++ Ani/ViewController/Initial/SignUpViewController/ANISignUpViewController.swift
@@ -1,23 +1,25 @@
 //
-//  ANILoginViewController.swift
+//  SignUpViewController.swift
 //  Ani
 //
-//  Created by jeonminseop on 2018/05/23.
+//  Created by jeonminseop on 2018/05/24.
 //  Copyright © 2018年 JeonMinseop. All rights reserved.
 //
 
 import UIKit
+import Gallery
 import TinyConstraints
 
-class ANILoginViewController: UIViewController {
-  
+class ANISignUpViewController: UIViewController {
   private weak var myNavigationBar: UIView?
   private weak var myNavigationBarBase: UIView?
   private weak var backButton: UIButton?
   
-  private var loginViewOriginalBottomConstraintConstant: CGFloat?
-  private var loginViewBottomConstraint: Constraint?
-  private weak var loginView: ANILoginView?
+  private var signUpViewOriginalBottomConstraintConstant: CGFloat?
+  private var signUpViewBottomConstraint: Constraint?
+  private weak var signUpView: ANISignUpView?
+  
+  private var gallery: GalleryController?
   
   private var rejectViewBottomConstraint: Constraint?
   private var rejectViewBottomConstraintOriginalConstant: CGFloat?
@@ -29,6 +31,10 @@ class ANILoginViewController: UIViewController {
   override func viewDidLoad() {
     setup()
     setupNotification()
+  }
+  
+  override func viewWillAppear(_ animated: Bool) {
+    UIApplication.shared.statusBar?.alpha = 1.0
   }
   
   private func setup() {
@@ -69,16 +75,16 @@ class ANILoginViewController: UIViewController {
     backButton.centerYToSuperview()
     self.backButton = backButton
     
-    //loginView
-    let loginView = ANILoginView()
-    loginView.delegate = self
-    self.view.addSubview(loginView)
-    loginView.topToBottom(of: myNavigationBar)
-    loginView.leftToSuperview()
-    loginView.rightToSuperview()
-    loginViewBottomConstraint = loginView.bottomToSuperview()
-    rejectViewBottomConstraintOriginalConstant = loginViewBottomConstraint?.constant
-    self.loginView = loginView
+    //signUpView
+    let signUpView = ANISignUpView()
+    signUpView.delegate = self
+    self.view.addSubview(signUpView)
+    signUpView.topToBottom(of: myNavigationBar)
+    signUpView.leftToSuperview()
+    signUpView.rightToSuperview()
+    signUpViewBottomConstraint = signUpView.bottomToSuperview()
+    signUpViewOriginalBottomConstraintConstant = signUpViewBottomConstraint?.constant
+    self.signUpView = signUpView
     
     //rejectView
     let rejectView = UIView()
@@ -104,7 +110,6 @@ class ANILoginViewController: UIViewController {
     rejectLabel.textAlignment = .center
     rejectLabel.textColor = .white
     rejectLabel.font = UIFont.boldSystemFont(ofSize: 16.0)
-    rejectLabel.text = "ユーザーが見つかりません！"
     rejectLabel.textAlignment = .center
     rejectBaseView.addSubview(rejectLabel)
     rejectLabel.edgesToSuperview()
@@ -120,11 +125,11 @@ class ANILoginViewController: UIViewController {
     guard let keyboardFrame = (notification.userInfo?[UIKeyboardFrameEndUserInfoKey] as? NSValue)?.cgRectValue,
           let duration = notification.userInfo?[UIKeyboardAnimationDurationUserInfoKey] as? TimeInterval,
           let curve = notification.userInfo?[UIKeyboardAnimationCurveUserInfoKey] as? UInt,
-          let loginViewBottomConstraint = self.loginViewBottomConstraint else { return }
+          let signUpViewBottomConstraint = self.signUpViewBottomConstraint else { return }
     
     let h = keyboardFrame.height
     
-    loginViewBottomConstraint.constant = -h
+    signUpViewBottomConstraint.constant = -h
     
     UIView.animate(withDuration: duration, delay: 0, options: UIViewAnimationOptions(rawValue: curve), animations: {
       self.view.layoutIfNeeded()
@@ -134,14 +139,33 @@ class ANILoginViewController: UIViewController {
   @objc private func keyboardWillHide(_ notification: Notification) {
     guard let duration = notification.userInfo?[UIKeyboardAnimationDurationUserInfoKey] as? TimeInterval,
           let curve = notification.userInfo?[UIKeyboardAnimationCurveUserInfoKey] as? UInt,
-          let signUpViewOriginalBottomConstraintConstant = self.loginViewOriginalBottomConstraintConstant,
-          let loginViewBottomConstraint = self.loginViewBottomConstraint else { return }
+          let signUpViewOriginalBottomConstraintConstant = self.signUpViewOriginalBottomConstraintConstant,
+          let signUpViewBottomConstraint = self.signUpViewBottomConstraint else { return }
     
-    loginViewBottomConstraint.constant = signUpViewOriginalBottomConstraintConstant
+    signUpViewBottomConstraint.constant = signUpViewOriginalBottomConstraintConstant
     
     UIView.animate(withDuration: duration, delay: 0, options: UIViewAnimationOptions(rawValue: curve), animations: {
       self.view.layoutIfNeeded()
     })
+  }
+  
+  func getCropImages(images: [UIImage?], items: [Image]) -> [UIImage] {
+    var croppedImages = [UIImage]()
+    
+    for (index, image) in images.enumerated() {
+      let imageSize = image?.size
+      let scrollViewWidth = self.view.frame.width
+      let widthScale =  scrollViewWidth / (imageSize?.width)! * items[index].scale
+      let heightScale = scrollViewWidth / (imageSize?.height)! * items[index].scale
+      
+      let scale = 1 / min(widthScale, heightScale)
+      let visibleRect = CGRect(x: items[index].offset.x * scale, y: items[index].offset.y * scale, width: scrollViewWidth * scale, height: scrollViewWidth * scale * Config.Grid.previewRatio)
+      let ref: CGImage = (image?.cgImage?.cropping(to: visibleRect))!
+      let croppedImage:UIImage = UIImage(cgImage: ref)
+      
+      croppedImages.append(croppedImage)
+    }
+    return croppedImages
   }
   
   //MARK: action
@@ -150,16 +174,41 @@ class ANILoginViewController: UIViewController {
   }
 }
 
-//MARK: ANILoginViewDelegate
-extension ANILoginViewController: ANILoginViewDelegate {
-  func loginSuccess() {
+//MARK: ANISignUpViewDelegate
+extension ANISignUpViewController: ANISignUpViewDelegate {
+  func signUpSuccess() {
     self.navigationController?.dismiss(animated: true, completion: nil)
+  }
+  
+  func donButtonTapped() {
+    self.navigationController?.dismiss(animated: true, completion: nil)
+  }
+  
+  func prifileImagePickButtonTapped() {
+    gallery = GalleryController()
+    if let galleryUnrap = gallery {
+      galleryUnrap.delegate = self
+      Gallery.Config.initialTab = .imageTab
+      Gallery.Config.PageIndicator.backgroundColor = .white
+      Gallery.Config.Camera.oneImageMode = true
+      if Gallery.Config.Camera.oneImageMode {
+        Gallery.Config.Grid.previewRatio = UIViewController.HEADER_IMAGE_VIEW_RATIO
+        Config.tabsToShow = [.imageTab, .cameraTab]
+      }
+      Gallery.Config.Font.Main.regular = UIFont.boldSystemFont(ofSize: 17)
+      Gallery.Config.Grid.ArrowButton.tintColor = ANIColor.dark
+      Gallery.Config.Grid.FrameView.borderColor = ANIColor.green
+      Gallery.Config.Grid.previewRatio = 1.0
+      
+      let galleryNV = UINavigationController(rootViewController: galleryUnrap)
+      self.present(galleryNV, animated: true, completion: nil)
+    }
   }
   
   func reject(notiText: String) {
     guard let rejectViewBottomConstraint = self.rejectViewBottomConstraint,
-          !isRejectAnimating,
-          let rejectLabel = self.rejectLabel else { return }
+          let rejectLabel = self.rejectLabel,
+          !isRejectAnimating else { return }
     
     rejectLabel.text = notiText
     
@@ -178,5 +227,47 @@ extension ANILoginViewController: ANILoginViewDelegate {
         self.isRejectAnimating = false
       })
     }
+  }
+}
+
+//MARK: GalleryControllerDelegate
+extension ANISignUpViewController: GalleryControllerDelegate {
+  func galleryController(_ controller: GalleryController, didSelectImages images: [Image]) {
+    
+    Image.resolve(images: images) { (myImages) in
+      let imageFilteriewController = ANIImageFilterViewController()
+      imageFilteriewController.images = self.getCropImages(images: myImages, items: images)
+      imageFilteriewController.delegate = self
+      controller.navigationController?.pushViewController(imageFilteriewController, animated: true)
+    }
+    
+    gallery = nil
+  }
+  
+  func galleryController(_ controller: GalleryController, didSelectVideo video: Video) {
+    controller.dismiss(animated: true, completion: nil)
+    
+    gallery = nil
+  }
+  
+  func galleryController(_ controller: GalleryController, requestLightbox images: [Image]) {
+    controller.dismiss(animated: true, completion: nil)
+    gallery = nil
+  }
+  
+  func galleryControllerDidCancel(_ controller: GalleryController) {
+    controller.dismiss(animated: true, completion: nil)
+    gallery = nil
+  }
+}
+
+//MARK: ANIImageFilterViewControllerDelegate
+extension ANISignUpViewController: ANIImageFilterViewControllerDelegate {
+  func doneFilterImages(filteredImages: [UIImage?]) {
+    guard !filteredImages.isEmpty,
+          let filteredImage = filteredImages[0],
+          let signUpView = self.signUpView else { return }
+    
+    signUpView.profileImage = filteredImage
   }
 }
